+++ packages/salesforcedx-vscode-core/src/commands/forceSourceDiff.ts
@@ -37,20 +37,22 @@ export async function forceSourceDiff(sourceUri?: vscode.Uri) {
   }
 
   const defaultUsernameorAlias = workspaceContext.username;
-  if (defaultUsernameorAlias) {
-    const executor = new MetadataCacheExecutor(
-      defaultUsernameorAlias,
-      nls.localize('force_source_diff_text'),
-      'force_source_diff',
-      handleCacheResults
-    );
-    const commandlet = new SfdxCommandlet(
-      workspaceChecker,
-      new FilePathGatherer(sourceUri),
-      executor
-    );
-    await commandlet.run();
+  if (!defaultUsernameorAlias) {
+    notificationService.showErrorMessage(nls.localize('missing_default_org'));
+    return;
   }
+  const executor = new MetadataCacheExecutor(
+    defaultUsernameorAlias,
+    nls.localize('force_source_diff_text'),
+    'force_source_diff',
+    handleCacheResults
+  );
+  const commandlet = new SfdxCommandlet(
+    workspaceChecker,
+    new FilePathGatherer(sourceUri),
+    executor
+  );
+  await commandlet.run();
 }
 
 export async function forceSourceFolderDiff(explorerPath: vscode.Uri) {