+++ packages/salesforcedx-vscode-core/src/commands/util/postconditionCheckers.ts
@@ -17,18 +17,44 @@ import {
   ConflictDetectionConfig,
   conflictDetector,
   conflictView,
-  DirectoryDiffResults
+  DirectoryDiffResults,
+  MetadataCacheService
 } from '../../conflict';
+import { TimestampConflictDetector } from '../../conflict/timestampConflictDetector';
 import { workspaceContext } from '../../context';
 import { nls } from '../../messages';
 import { notificationService } from '../../notifications';
-import { sfdxCoreSettings } from '../../settings';
+import { DeployQueue, sfdxCoreSettings } from '../../settings';
 import { telemetryService } from '../../telemetry';
 import { getRootWorkspacePath, MetadataDictionary } from '../../util';
 import { PathStrategyFactory } from './sourcePathStrategies';
 
 type OneOrMany = LocalComponent | LocalComponent[];
 type ContinueOrCancel = ContinueResponse<OneOrMany> | CancelResponse;
+
+export class CompositePostconditionChecker<T>
+  implements PostconditionChecker<T> {
+  private readonly postcheckers: Array<PostconditionChecker<any>>;
+  public constructor(...postcheckers: Array<PostconditionChecker<any>>) {
+    this.postcheckers = postcheckers;
+  }
+  public async check(
+    inputs: CancelResponse | ContinueResponse<T>
+  ): Promise<CancelResponse | ContinueResponse<T>> {
+    if (inputs.type === 'CONTINUE') {
+      const aggregatedData: any = {};
+      for (const postchecker of this.postcheckers) {
+        inputs = await postchecker.check(inputs);
+        if (inputs.type !== 'CONTINUE') {
+          return {
+            type: 'CANCEL'
+          };
+        }
+      }
+    }
+    return inputs;
+  }
+}
 
 export class EmptyPostChecker implements PostconditionChecker<any> {
   public async check(
@@ -244,7 +270,7 @@ export class ConflictDetectionChecker implements PostconditionChecker<string> {
         )
       );
       results.different.forEach(file => {
-        channelService.appendLine(normalize(file));
+        channelService.appendLine(normalize(file.path));
       });
       channelService.showChannelOutput();
 
@@ -282,5 +308,130 @@ export class ConflictDetectionChecker implements PostconditionChecker<string> {
       }
     }
     return { type: 'CONTINUE', data: manifest };
+  }
+}
+
+export class TimestampConflictChecker implements PostconditionChecker<string> {
+  private isManifest: boolean;
+  private messages: ConflictDetectionMessages;
+
+  constructor(isManifest: boolean, messages: ConflictDetectionMessages) {
+    this.messages = messages;
+    this.isManifest = isManifest;
+  }
+
+  public async check(
+    inputs: ContinueResponse<string> | CancelResponse
+  ): Promise<ContinueResponse<string> | CancelResponse> {
+    if (!sfdxCoreSettings.getConflictDetectionEnabled()) {
+      return inputs;
+    }
+
+    if (inputs.type === 'CONTINUE') {
+      channelService.showChannelOutput();
+      channelService.showCommandWithTimestamp(
+        `${nls.localize('channel_starting_message')}${nls.localize(
+          'conflict_detect_execution_name'
+        )}\n`
+      );
+
+      const { username } = workspaceContext;
+      if (!username) {
+        return {
+          type: 'CANCEL',
+          msg: nls.localize('conflict_detect_no_default_username')
+        };
+      }
+
+      const componentPath = inputs.data;
+      const cacheService = new MetadataCacheService(username);
+
+      try {
+        const result = await cacheService.loadCache(
+          componentPath,
+          getRootWorkspacePath(),
+          this.isManifest
+        );
+        const detector = new TimestampConflictDetector();
+        const diffs = detector.createDiffs(result);
+
+        channelService.showCommandWithTimestamp(
+          `${nls.localize('channel_end')} ${nls.localize(
+            'conflict_detect_execution_name'
+          )}\n`
+        );
+        return await this.handleConflicts(inputs.data, username, diffs);
+      } catch (error) {
+        console.error(error);
+        const errorMsg = nls.localize(
+          'conflict_detect_error',
+          error.toString()
+        );
+        channelService.appendLine(errorMsg);
+        telemetryService.sendException('ConflictDetectionException', errorMsg);
+        await DeployQueue.get().unlock();
+      }
+    }
+    return { type: 'CANCEL' };
+  }
+
+  public async handleConflicts(
+    componentPath: string,
+    usernameOrAlias: string,
+    results: DirectoryDiffResults
+  ): Promise<ContinueResponse<string> | CancelResponse> {
+    const conflictTitle = nls.localize(
+      'conflict_detect_view_root',
+      usernameOrAlias,
+      results.different.size
+    );
+
+    if (results.different.size === 0) {
+      conflictView.visualizeDifferences(conflictTitle, usernameOrAlias, false);
+    } else {
+      channelService.appendLine(
+        nls.localize(
+          'conflict_detect_conflict_header_timestamp',
+          results.different.size
+        )
+      );
+      results.different.forEach(file => {
+        channelService.appendLine(normalize(file.path));
+      });
+
+      const choice = await notificationService.showWarningModal(
+        nls.localize(this.messages.warningMessageKey),
+        nls.localize('conflict_detect_show_conflicts'),
+        nls.localize('conflict_detect_override')
+      );
+
+      if (choice === nls.localize('conflict_detect_override')) {
+        conflictView.visualizeDifferences(
+          conflictTitle,
+          usernameOrAlias,
+          false
+        );
+      } else {
+        channelService.appendLine(
+          nls.localize(
+            'conflict_detect_command_hint',
+            this.messages.commandHint(componentPath)
+          )
+        );
+
+        const doReveal =
+          choice === nls.localize('conflict_detect_show_conflicts');
+        conflictView.visualizeDifferences(
+          conflictTitle,
+          usernameOrAlias,
+          doReveal,
+          results
+        );
+
+        await DeployQueue.get().unlock();
+        return { type: 'CANCEL' };
+      }
+    }
+    return { type: 'CONTINUE', data: componentPath };
   }
 }