+++ packages/salesforcedx-vscode-core/test/vscode-integration/commands/functions/forceFunctionStop.test.ts
@@ -4,25 +4,25 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { CliCommandExecutor } from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
+import * as library from '@heroku/functions-core';
 import * as path from 'path';
 import { assert, createSandbox, SinonSandbox, SinonStub } from 'sinon';
+import Sinon = require('sinon');
 import { Uri } from 'vscode';
 import { channelService } from '../../../../src/channels';
-import {
-  forceFunctionStart,
-  ForceFunctionStartExecutor
-} from '../../../../src/commands/functions/forceFunctionStart';
+import { forceFunctionStart } from '../../../../src/commands/functions/forceFunctionStart';
 import { forceFunctionStop } from '../../../../src/commands/functions/forceFunctionStop';
+import { FunctionService } from '../../../../src/commands/functions/functionService';
 import { nls } from '../../../../src/messages';
 import { notificationService } from '../../../../src/notifications';
 import { telemetryService } from '../../../../src/telemetry';
 import { getRootWorkspacePath } from '../../../../src/util';
-import { MockExecution } from './mockExecution';
 
 describe('Force Function Stop', () => {
   let sandbox: SinonSandbox;
-  let cliCommandExecutorStub: SinonStub;
+  const functionsBinaryStub: {
+    [key: string]: SinonStub;
+  } = {};
   const channelServiceStubs: {
     [key: string]: SinonStub;
   } = {};
@@ -32,14 +32,14 @@ describe('Force Function Stop', () => {
   const telemetryServiceStubs: {
     [key: string]: SinonStub;
   } = {};
+  const functionServiceStubs: {
+    [key: string]: SinonStub;
+  } = {};
   let hrtimeStub: SinonStub;
   beforeEach(() => {
     sandbox = createSandbox();
-
-    cliCommandExecutorStub = sandbox.stub(
-      CliCommandExecutor.prototype,
-      'execute'
-    );
+    functionsBinaryStub.cancel = sandbox.stub();
+    sandbox.stub(library, 'getFunctionsBinary').returns(functionsBinaryStub);
     channelServiceStubs.appendLineStub = sandbox.stub(
       channelService,
       'appendLine'
@@ -67,21 +67,24 @@ describe('Force Function Stop', () => {
   });
 
   it('Should stop function, show notification and send telemetry', async () => {
+    const FUNCTION_LANGUAGE = 'node';
+    functionServiceStubs.getFunctionLanguage = sandbox.stub(
+      FunctionService.prototype,
+      'getFunctionLanguage'
+    );
+    functionServiceStubs.getFunctionLanguage.returns(FUNCTION_LANGUAGE);
     const srcUri = Uri.file(
       path.join(getRootWorkspacePath(), 'functions', 'demoJavaScriptFunction')
     );
-    const executor = new ForceFunctionStartExecutor();
-    const mockExecution = new MockExecution(executor.build(srcUri.fsPath));
-    const killExecutionStub = sandbox.stub(mockExecution, 'killExecution');
-    cliCommandExecutorStub.returns(mockExecution);
+
     await forceFunctionStart(srcUri);
 
     const mockStartTime = [1234, 5678];
     hrtimeStub.returns(mockStartTime);
     await forceFunctionStop();
 
-    assert.calledOnce(killExecutionStub);
-    assert.calledOnce(channelServiceStubs.appendLineStub);
+    assert.calledOnce(functionsBinaryStub.cancel);
+    assert.called(channelServiceStubs.appendLineStub);
     assert.calledWith(
       channelServiceStubs.appendLineStub,
       nls.localize('force_function_stop_in_progress')
@@ -95,7 +98,8 @@ describe('Force Function Stop', () => {
     assert.calledWith(
       telemetryServiceStubs.sendCommandEventStub,
       'force_function_stop',
-      mockStartTime
+      mockStartTime,
+      { language: FUNCTION_LANGUAGE }
     );
   });
 
@@ -113,9 +117,7 @@ describe('Force Function Stop', () => {
     const srcUri = Uri.file(
       path.join(getRootWorkspacePath(), 'functions', 'demoJavaScriptFunction')
     );
-    const executor = new ForceFunctionStartExecutor();
-    const mockExecution = new MockExecution(executor.build(srcUri.fsPath));
-    cliCommandExecutorStub.returns(mockExecution);
+
     await forceFunctionStart(srcUri);
 
     const mockStartTime = [1234, 5678];